Refactor test code
- move functions that support tests from mob_test.go into new mob_testutils.go
- trim mob_test.go imports
