+++ mob_test.go
@@ -1,13 +1,7 @@
 package main
 
 import (
-	fmt "fmt"
-	"io/ioutil"
 	"os"
-	"path/filepath"
-	"reflect"
-	"runtime"
-	"strconv"
 	"strings"
 	"testing"
 )
@@ -552,137 +546,4 @@ func TestDoneMerge(t *testing.T) {
 	start()
 	done()
 	assertOutputContains(t, output, "git commit -m 'describe the changes'")
-}
-
-func setup(t *testing.T) *string {
-	configuration = getDefaultConfiguration()
-	output := captureOutput()
-	createTestbed(t)
-	assertOnBranch(t, "master")
-	equals(t, "master", gitBranches())
-	equals(t, "origin/master", gitRemoteBranches())
-	assertNoMobSessionBranches(t, "mob-session")
-	return output
-}
-
-func captureOutput() *string {
-	messages := ""
-	printToConsole = func(text string) {
-		fmt.Print(text)
-		messages += text
-	}
-	return &messages
-}
-
-func run(t *testing.T, name string, args ...string) {
-	commandString, output, err := runCommand(name, args...)
-	if err != nil {
-		fmt.Println(commandString)
-		fmt.Println(output)
-		fmt.Println(err.Error())
-		t.Error("command " + commandString + " failed")
-	}
-}
-
-func createTestbed(t *testing.T) {
-	workingDir = ""
-	run(t, "./create-testbed")
-
-	setWorkingDir("/tmp/mob/local")
-	assertOnBranch(t, "master")
-	assertNoMobSessionBranches(t, "mob-session")
-}
-
-func setWorkingDir(dir string) {
-	workingDir = dir
-	say("\nSET WORKING DIR TO " + dir + "\n======================\n")
-}
-
-func assertCommits(t *testing.T, commits int) {
-	result := silentgit("rev-list", "--count", "HEAD")
-	number, _ := strconv.Atoi(strings.TrimSpace(result))
-	if number != commits {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, strconv.Itoa(commits)+" commits in "+workingDir, strconv.Itoa(number)+" commits in "+workingDir)
-		t.FailNow()
-	}
-}
-
-func assertFileExist(t *testing.T, filename string) {
-	path := workingDir + "/" + filename
-	if _, err := os.Stat(path); os.IsNotExist(err) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "existing file "+path, "no file at "+path)
-		t.FailNow()
-	}
-}
-
-func createFile(t *testing.T, filename string, content string) {
-	d1 := []byte(content)
-	err := ioutil.WriteFile(workingDir+"/"+filename, d1, 0644)
-	if err != nil {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "creating file "+filename+" with content "+content, "error")
-		t.FailNow()
-	}
-}
-
-func assertOnBranch(t *testing.T, branch string) {
-	currentBranch := gitCurrentBranch()
-	if currentBranch != branch {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "on branch "+branch, "on branch "+currentBranch)
-		t.FailNow()
-	}
-}
-
-func assertOutputContains(t *testing.T, output *string, contains string) {
-	currentOutput := *output
-	if !strings.Contains(currentOutput, contains) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "output contains '"+contains+"'", currentOutput)
-		t.FailNow()
-	}
-}
-
-func assertOutputNotContains(t *testing.T, output *string, notContains string) {
-	if strings.Contains(*output, notContains) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "output not contains "+notContains, output)
-		t.FailNow()
-	}
-}
-
-func assertMobSessionBranches(t *testing.T, branch string) {
-	if !hasRemoteBranch(branch) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, configuration.RemoteName+"/"+branch, "none")
-		t.FailNow()
-	}
-	if !hasLocalBranch(branch) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, branch, "none")
-		t.FailNow()
-	}
-}
-
-func assertNoMobSessionBranches(t *testing.T, branch string) {
-	if hasRemoteBranch(branch) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "none", configuration.RemoteName+"/"+branch)
-		t.FailNow()
-	}
-	if hasLocalBranch(branch) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, "none", branch)
-		t.FailNow()
-	}
-}
-
-func equals(t *testing.T, exp, act interface{}) {
-	if !reflect.DeepEqual(exp, act) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, exp, act)
-		t.FailNow()
-	}
 }
